# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LoveMediator API
  version: 1.0.0
  description: |
    OpenAPI standard file generated from `docs/API_LoveMediator_v1.md`.
    All path identifiers such as `eventId`, `reviewId`, `relationshipId`, and `userId`
    refer to public IDs rather than database primary keys.
servers:
  - url: /api/v1
    description: API v1
...
security:
  - BearerAuth: []
paths:
...
  /relationships:
    get:
      tags: [relationships]
      summary: List active relationships for current user
      operationId: listRelationships
      responses:
        '200':
          description: Relationship list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipListResponseEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /relationships/invite:
    post:
      tags: [relationships]
      summary: Create relationship invite
      operationId: createRelationshipInvite
      responses:
        '200':
          description: Relationship invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipInviteResponseEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /relationships/accept:
    post:
      tags: [relationships]
      summary: Accept relationship invite
      operationId: acceptRelationshipInvite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationshipAcceptRequest'
      responses:
        '200':
          description: Relationship accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipAcceptResponseEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /relationships/{relationshipId}/analysis-sessions/a:
    post:
      tags: [relationships]
      summary: Start or resume A private analysis session
      operationId: startAAnalysisSession
      parameters:
        - $ref: '#/components/parameters/RelationshipId'
      responses:
        '200':
          description: Analysis session ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisSessionResponseEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /relationships/{relationshipId}/cancel-request:
    post:
      tags: [relationships]
      summary: Create relationship cancel confirmation token
      operationId: createRelationshipCancelRequest
      parameters:
        - $ref: '#/components/parameters/RelationshipId'
      responses:
        '200':
          description: Cancel confirmation token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipCancelRequestResponseEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /relationships/{relationshipId}/cancel-confirm:
    post:
      tags: [relationships]
      summary: Confirm relationship cancellation and hard-delete related data
      operationId: confirmRelationshipCancel
      parameters:
        - $ref: '#/components/parameters/RelationshipId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationshipCancelConfirmRequest'
      responses:
        '200':
          description: Relationship deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipCancelResponseEnvelope'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RelationshipId:
      in: path
      name: relationshipId
      required: true
      description: Relationship public ID
      schema:
        type: string
        example: rel_001
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
          examples:
            unauthorized:
              value:
                code: 2001
                message: 用户名或密码错误
                data: null
  schemas:
    EventStatus:
      type: string
      enum: [draft, waiting_b, judged, reviewed, closed]
...
    ApiEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          description: Response payload
          nullable: true
    ErrorResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            code:
              type: integer
              example: 2001
            message:
              type: string
              example: 用户名或密码错误
            data:
              nullable: true
              example: null
...
    RelationshipCurrentEvent:
      type: object
      required: [eventId, status, pendingAction]
      properties:
        eventId:
          type: string
          example: ev_001
        status:
          $ref: '#/components/schemas/EventStatus'
        pendingAction:
          type: string
          example: respond
        title:
          type: string
          nullable: true
          example: null
    RelationshipSummary:
      type: object
      required: [relationshipId, partnerUserId, partnerUsername, status]
      properties:
        relationshipId:
          type: string
          example: rel_001
        partnerUserId:
          type: string
          example: u_1002
        partnerUsername:
          type: string
          example: bob_01
        status:
          type: string
          enum: [active]
        currentEvent:
          allOf:
            - $ref: '#/components/schemas/RelationshipCurrentEvent'
          nullable: true
    RelationshipListData:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RelationshipSummary'
    RelationshipListResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RelationshipListData'
    RelationshipInviteData:
      type: object
      required: [inviteToken, inviteUrl, expiresAt]
      properties:
        inviteToken:
          type: string
          example: invite_token_xxx
        inviteUrl:
          type: string
          example: /relationship-invite?token=invite_token_xxx
        expiresAt:
          type: string
          format: date-time
          example: '2026-04-03T08:21:13Z'
    RelationshipInviteResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RelationshipInviteData'
    RelationshipAcceptRequest:
      type: object
      required: [inviteToken]
      properties:
        inviteToken:
          type: string
          example: invite_token_xxx
    RelationshipAcceptData:
      allOf:
        - $ref: '#/components/schemas/RelationshipSummary'
    RelationshipAcceptResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RelationshipAcceptData'
    RelationshipCancelConfirmRequest:
      type: object
      required: [cancelToken]
      properties:
        cancelToken:
          type: string
          example: cancel_token_xxx
    RelationshipCancelRequestData:
      type: object
      required: [cancelToken, cancelUrl, expiresAt]
      properties:
        cancelToken:
          type: string
          example: cancel_token_xxx
        cancelUrl:
          type: string
          example: /relationship-cancel?token=cancel_token_xxx
        expiresAt:
          type: string
          format: date-time
          example: '2026-04-03T08:21:13Z'
    RelationshipCancelRequestResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RelationshipCancelRequestData'
    RelationshipDeleteCounts:
      type: object
      required:
        [privateMessages, privateSessions, followupMessages, eventStateLogs, aiCallLogs, elfMessages, judgeResults, eventSnapshots, calendarEntries, reviewVersions, reviews, events, relationships]
      properties:
        privateMessages:
          type: integer
        privateSessions:
          type: integer
        followupMessages:
          type: integer
        eventStateLogs:
          type: integer
        aiCallLogs:
          type: integer
        elfMessages:
          type: integer
        judgeResults:
          type: integer
        eventSnapshots:
          type: integer
        calendarEntries:
          type: integer
        reviewVersions:
          type: integer
        reviews:
          type: integer
        events:
          type: integer
        relationships:
          type: integer
    RelationshipCancelData:
      type: object
      required: [relationshipId, deletedCounts, deletedAt]
      properties:
        relationshipId:
          type: string
          example: rel_001
        deletedCounts:
          $ref: '#/components/schemas/RelationshipDeleteCounts'
        deletedAt:
          type: string
          format: date-time
          example: '2026-03-27T10:00:00Z'
    RelationshipCancelResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RelationshipCancelData'
    AnalysisSessionMessagePayload:
      type: object
      required: [role, content, createdAt]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
          example: '2026-03-27T10:00:00Z'
    AnalysisSessionData:
      type: object
      required: [sessionId, phase, expiresAt, messages]
      properties:
        sessionId:
          type: string
          example: as_001
        phase:
          type: string
          enum: [a, b]
        relationshipId:
          type: string
          nullable: true
          example: rel_001
        eventId:
          type: string
          nullable: true
          example: ev_001
        expiresAt:
          type: string
          format: date-time
          example: '2026-03-28T10:00:00Z'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisSessionMessagePayload'
    AnalysisSessionResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AnalysisSessionData'
...